refactor(cli-dev): extract remaining test utils
