+++ packages/cli/commons/src/preconditions/mockPreconditions.ts
@@ -1,0 +1,27 @@
+import {Command} from '@oclif/core';
+import {PreconditionError} from '../errors/preconditionError';
+import {PreconditionFunction} from './preconditions';
+
+const thrower = (reason: string) => {
+  throw new PreconditionError(`${reason} Precondition Error`);
+};
+
+export const mockPreconditions = <
+  PreconditionStatus extends Record<string, boolean>
+>(
+  preconditionStatus: PreconditionStatus
+) => {
+  type PreconditionKeys = keyof PreconditionStatus & string;
+  type PreconditionPromises = Record<PreconditionKeys, PreconditionFunction>;
+
+  const keys: PreconditionKeys[] = Object.keys(preconditionStatus);
+  const mockedPreconditions: Partial<PreconditionPromises> = {};
+  for (const key of keys) {
+    mockedPreconditions[key] = (_target?: Command) =>
+      new Promise<void>((resolve) =>
+        preconditionStatus[key] ? resolve() : thrower(key)
+      );
+  }
+
+  return mockedPreconditions as PreconditionPromises;
+};
