+++ packages/cli/core/src/lib/decorators/preconditions/node.spec.ts
@@ -2,7 +2,7 @@ jest.mock('../../utils/process');
 
 import {dedent} from 'ts-dedent';
 import {spawnProcessOutput} from '../../utils/process';
-import {getFakeCommand} from './testsUtils/utils';
+import {getFakeCommand} from '@coveo/cli-commons-dev/testUtils/utils';
 
 import {IsNodeVersionInRange} from './node';
 import {fancyIt} from '@coveo/cli-commons-dev/testUtils/it';
