+++ packages/cli/core/src/lib/snapshot/expandedPreviewer/filesDiffProcessor.spec.ts
@@ -3,7 +3,7 @@ jest.mock('fs-extra');
 
 import {readdirSync, rmSync} from 'fs';
 import {readJSONSync, writeJSONSync} from 'fs-extra';
-import {getDirectory, getFile} from '../../../__test__/fsUtils';
+import {getDirectory, getFile} from '@coveo/cli-commons-dev/testUtils/fsUtils';
 import {recursiveDirectoryDiff} from './filesDiffProcessor';
 import {join} from 'path';
 import {fancyIt} from '@coveo/cli-commons-dev/testUtils/it';
